CCCP-227, refactor: move events into rbclib/events
