+++ rbclib/periodic/vsp_feed.py
@@ -11,7 +11,7 @@ from chainpy.logger import global_logger
 from rbclib.metric import PrometheusExporterRelayer
 from rbclib.primitives.consts import SOCKET_CONTRACT_NAME, ROUND_UP_FUNCTION_NAME, NoneParams
 from rbclib.primitives.relay_chain import chain_enum
-from rbclib.relayersubmit import SocketSignature
+from rbclib.submits import SocketSignature
 from rbclib.utils import fetch_bottom_round, fetch_latest_round, fetch_relayer_index, is_selected_relayer, fetch_sorted_relayer_list_lower, \
     log_invalid_flow
 from relayer.global_config import relayer_config_global
